# docs/swagger.yaml
basePath: /
definitions:
  models.AnnualReportIncomeStatements:
    properties:
      comprehensiveIncomeNetOfTax:
        type: integer
      costOfRevenue:
        type: integer
      costofGoodsAndServicesSold:
        type: integer
      depreciation:
        type: integer
      depreciationAndAmortization:
        type: integer
      ebit:
        type: integer
      ebitda:
        type: integer
      fiscalDateEnding:
        type: string
      grossProfit:
        type: integer
      incomeBeforeTax:
        type: integer
      incomeTaxExpense:
        type: integer
      interestAndDebtExpense:
        type: integer
      interestExpense:
        type: integer
      interestIncome:
        type: integer
      investmentIncomeNet:
        description: Have to handle case where value is "None"
        type: integer
      netIncome:
        type: integer
      netIncomeFromContinuingOperations:
        type: integer
      netInterestIncome:
        type: integer
      nonInterestIncome:
        type: integer
      operatingExpenses:
        type: integer
      operatingIncome:
        type: integer
      otherNonOperatingIncome:
        type: integer
      reportedCurrency:
        type: string
      researchAndDevelopment:
        type: integer
      sellingGeneralAndAdministrative:
        type: integer
      totalRevenue:
        type: integer
    type: object
  models.BalanceSheetAnnualReport:
    properties:
      accumulatedDepreciationAmortizationPPE:
        type: integer
      capitalLeaseObligations:
        type: integer
      cashAndCashEquivalentsAtCarryingValue:
        type: integer
      cashAndShortTermInvestments:
        type: integer
      commonStock:
        type: integer
      commonStockSharesOutstanding:
        type: integer
      currentAccountsPayable:
        type: integer
      currentDebt:
        type: integer
      currentLongTermDebt:
        type: integer
      currentNetReceivables:
        type: integer
      deferredRevenue:
        type: integer
      fiscalDateEnding:
        type: string
      goodwill:
        type: integer
      intangibleAssets:
        type: integer
      intangibleAssetsExcludingGoodwill:
        type: integer
      inventory:
        type: integer
      investments:
        type: integer
      longTermDebt:
        type: integer
      longTermDebtNoncurrent:
        type: integer
      longTermInvestments:
        type: integer
      otherCurrentAssets:
        type: integer
      otherCurrentLiabilities:
        type: integer
      otherNonCurrentAssets:
        type: integer
      otherNonCurrentLiabilities:
        type: integer
      propertyPlantEquipment:
        type: integer
      reportedCurrency:
        type: string
      retainedEarnings:
        type: integer
      shortLongTermDebtTotal:
        type: integer
      shortTermDebt:
        type: integer
      shortTermInvestments:
        type: integer
      totalAssets:
        type: integer
      totalCurrentAssets:
        type: integer
      totalCurrentLiabilities:
        type: integer
      totalLiabilities:
        type: integer
      totalNonCurrentAssets:
        type: integer
      totalNonCurrentLiabilities:
        type: integer
      totalShareholderEquity:
        type: integer
      treasuryStock:
        type: integer
    type: object
  models.CashFlowStatementsAnnualReport:
    properties:
      capitalExpenditures:
        type: integer
      cashflowFromFinancing:
        type: integer
      cashflowFromInvestment:
        type: integer
      changeInCashAndCashEquivalents:
        type: integer
      changeInExchangeRate:
        type: integer
      changeInInventory:
        type: integer
      changeInOperatingAssets:
        type: integer
      changeInOperatingLiabilities:
        type: integer
      changeInReceivables:
        type: integer
      depreciationDepletionAndAmortization:
        type: integer
      dividendPayout:
        type: integer
      dividendPayoutCommonStock:
        type: integer
      dividendPayoutPreferredStock:
        type: integer
      fiscalDateEnding:
        type: string
      netIncome:
        type: integer
      operatingCashflow:
        type: integer
      paymentsForOperatingActivities:
        type: integer
      paymentsForRepurchaseOfCommonStock:
        type: integer
      paymentsForRepurchaseOfEquity:
        type: integer
      paymentsForRepurchaseOfPreferredStock:
        type: integer
      proceedsFromIssuanceOfCommonStock:
        type: integer
      proceedsFromIssuanceOfLongTermDebtAndCapitalSecuritiesNet:
        type: integer
      proceedsFromIssuanceOfPreferredStock:
        type: integer
      proceedsFromOperatingActivities:
        type: integer
      proceedsFromRepaymentsOfShortTermDebt:
        type: integer
      proceedsFromRepurchaseOfEquity:
        type: integer
      proceedsFromSaleOfTreasuryStock:
        type: integer
      profitLoss:
        type: integer
      reportedCurrency:
        type: string
    type: object
  models.Company:
    properties:
      country:
        type: string
      currency:
        type: string
      exchange:
        type: string
      id:
        type: integer
      industry:
        type: string
      name:
        type: string
      sector:
        type: string
      symbol:
        type: string
    type: object
  models.CompanyAnalysis:
    properties:
      calculated_ratios:
        allOf:
        - $ref: '#/definitions/models.FinancialRatios'
        description: Ratios calculés
      company:
        $ref: '#/definitions/models.Company'
      growth_metrics:
        allOf:
        - $ref: '#/definitions/models.GrowthMetrics'
        description: Métriques dérivées
      latest_balance:
        $ref: '#/definitions/models.BalanceSheetAnnualReport'
      latest_cash_flow:
        $ref: '#/definitions/models.CashFlowStatementsAnnualReport'
      latest_income:
        $ref: '#/definitions/models.AnnualReportIncomeStatements'
      overview:
        $ref: '#/definitions/models.CompanyOverview'
    type: object
  models.CompanyOverview:
    properties:
      50DayMovingAverage:
        example: "0"
        type: string
      52WeekHigh:
        example: "0"
        type: string
      52WeekLow:
        example: "0"
        type: string
      200DayMovingAverage:
        example: "0"
        type: string
      Address:
        type: string
      AnalystTargetPrice:
        example: "0"
        type: string
      AssetType:
        type: string
      Beta:
        example: "0"
        type: string
      BookValue:
        example: "0"
        type: string
      CIK:
        type: string
      Country:
        type: string
      Currency:
        type: string
      Description:
        type: string
      DilutedEPSTTM:
        example: "0"
        type: string
      DividendDate:
        description: '"2023-06-10"'
        type: string
      DividendPerShare:
        example: "0"
        type: string
      DividendYield:
        example: "0"
        type: string
      EBITDA:
        example: "0"
        type: string
      EPS:
        example: "0"
        type: string
      EVToEBITDA:
        example: "0"
        type: string
      EVToRevenue:
        example: "0"
        type: string
      ExDividendDate:
        description: '"2023-05-09"'
        type: string
      Exchange:
        type: string
      FiscalYearEnd:
        type: string
      ForwardPE:
        example: "0"
        type: string
      GrossProfitTTM:
        example: "0"
        type: string
      Industry:
        type: string
      LatestQuarter:
        description: '"2023-03-31"'
        type: string
      MarketCapitalization:
        example: "0"
        type: string
      Name:
        type: string
      OperatingMarginTTM:
        example: "0"
        type: string
      PEGRatio:
        example: "0"
        type: string
      PERatio:
        example: "0"
        type: string
      PriceToBookRatio:
        example: "0"
        type: string
      PriceToSalesRatioTTM:
        example: "0"
        type: string
      ProfitMargin:
        example: "0"
        type: string
      QuarterlyEarningsGrowthYOY:
        example: "0"
        type: string
      QuarterlyRevenueGrowthYOY:
        example: "0"
        type: string
      ReturnOnAssetsTTM:
        example: "0"
        type: string
      ReturnOnEquityTTM:
        example: "0"
        type: string
      RevenuePerShareTTM:
        example: "0"
        type: string
      RevenueTTM:
        example: "0"
        type: string
      Sector:
        type: string
      SharesOutstanding:
        example: "0"
        type: string
      Symbol:
        type: string
      TrailingPE:
        example: "0"
        type: string
    type: object
  models.FairValue:
    properties:
      estimated_value:
        type: number
      method:
        type: string
      upside:
        type: number
    type: object
  models.FinancialRatios:
    properties:
      asset_turnover:
        type: number
      debt_to_assets:
        type: number
      debt_to_equity:
        type: number
      earnings_growth_yoy:
        type: number
      fiscal_date:
        type: string
      gross_margin:
        type: number
      inventory_turnover:
        type: number
      net_margin:
        type: number
      operating_margin:
        type: number
      pb_ratio:
        type: number
      pe_ratio:
        type: number
      price_to_sales_ratio:
        type: number
      revenue_growth_yoy:
        type: number
    type: object
  models.GrowthMetrics:
    properties:
      earnings_growth_1y:
        type: number
      earnings_growth_3y:
        type: number
      revenue_growth_1y:
        type: number
      revenue_growth_3y:
        type: number
    type: object
  models.HistoricalRatios:
    properties:
      fiscal_date:
        type: string
      ratios:
        $ref: '#/definitions/models.ValuationRatios'
    type: object
  models.PeerCompany:
    properties:
      market_cap:
        type: integer
      name:
        type: string
      pb_ratio:
        type: number
      pe_ratio:
        type: number
      symbol:
        type: string
    type: object
  models.ScreeningResult:
    properties:
      beta:
        type: number
      company_id:
        type: integer
      industry:
        type: string
      market_cap:
        type: integer
      name:
        type: string
      pb_ratio:
        type: number
      pe_ratio:
        type: number
      price_to_sales:
        type: number
      profit_margin:
        type: number
      roa:
        type: number
      roe:
        type: number
      score:
        description: Score calculé basé sur les critères
        type: number
      sector:
        type: string
      symbol:
        type: string
    type: object
  models.SectorAverages:
    properties:
      avg_pb_ratio:
        type: number
      avg_pe_ratio:
        type: number
      avg_profit_margin:
        type: number
      avg_revenue_growth:
        type: number
      avg_roe:
        type: number
      company_count:
        type: integer
    type: object
  models.SectorComparison:
    properties:
      company:
        $ref: '#/definitions/models.Company'
      peer_companies:
        items:
          $ref: '#/definitions/models.PeerCompany'
        type: array
      sector_averages:
        $ref: '#/definitions/models.SectorAverages'
    type: object
  models.ValuationAnalysis:
    properties:
      current_ratios:
        $ref: '#/definitions/models.ValuationRatios'
      fair_value:
        $ref: '#/definitions/models.FairValue'
      historical_ratios:
        items:
          $ref: '#/definitions/models.HistoricalRatios'
        type: array
      symbol:
        type: string
    type: object
  models.ValuationRatios:
    properties:
      enterprise_value_ebit:
        type: number
      pb_ratio:
        type: number
      pe_ratio:
        type: number
      price_to_cash_flow:
        type: number
      price_to_sales:
        type: number
    type: object
  pkg.CompanyAnalysisResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.CompanyAnalysis'
        description: Analyse complète de l'entreprise
      status:
        description: Statut de la réponse
        type: string
    type: object
  pkg.ErrorResponse:
    properties:
      code:
        description: Code d'erreur HTTP
        type: integer
      message:
        description: Message d'erreur
        type: string
      status:
        description: Statut de la réponse, e.g., "error"
        type: string
    type: object
  pkg.ScreeningResponse:
    properties:
      count:
        description: Nombre de résultats
        type: integer
      data:
        description: Résultats du screening
        items:
          $ref: '#/definitions/models.ScreeningResult'
        type: array
      status:
        description: Statut de la réponse
        type: string
    type: object
  pkg.ScreeningTemplate:
    properties:
      criteria:
        description: Critères du template
      description:
        description: Description
        type: string
      name:
        description: Nom du template
        type: string
    type: object
  pkg.ScreeningTemplatesResponse:
    properties:
      status:
        description: Statut de la réponse
        type: string
      templates:
        description: Liste des templates
        items:
          $ref: '#/definitions/pkg.ScreeningTemplate'
        type: array
    type: object
  pkg.SectorComparisonResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.SectorComparison'
        description: Comparaison sectorielle
      status:
        description: Statut de la réponse
        type: string
    type: object
  pkg.SuccessResponse:
    properties:
      message:
        description: Message de succès
        type: string
      status:
        description: Statut de la réponse
        type: string
    type: object
  pkg.ValuationAnalysisResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/models.ValuationAnalysis'
        description: Analyse de valorisation
      status:
        description: Statut de la réponse
        type: string
    type: object
  services.ScreeningCriteria:
    properties:
      limit:
        description: Limit results
        type: integer
      max_debt_to_equity:
        description: Financial health
        type: number
      max_market_cap:
        type: integer
      max_pb_ratio:
        type: number
      max_pe_ratio:
        type: number
      min_current_ratio:
        type: number
      min_eps_growth:
        type: number
      min_market_cap:
        description: Size
        type: integer
      min_pb_ratio:
        type: number
      min_pe_ratio:
        description: Valuation ratios
        type: number
      min_price_to_sales:
        type: number
      min_profit_margin:
        type: number
      min_revenue_growth:
        description: Growth
        type: number
      min_roa:
        type: number
      min_roe:
        description: Profitability
        type: number
      sectors:
        description: Sectors (optionnel)
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: API for analyzing company financials based on Alpha Vantage data.
  title: My Investment API
  version: "1.0"
paths:
  /api/v1/analysis/{symbol}:
    get:
      consumes:
      - application/json
      description: Get comprehensive fundamental analysis for a specific company
      parameters:
      - description: Company Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.CompanyAnalysisResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Get detailed company analysis
      tags:
      - Analysis
  /api/v1/analysis/{symbol}/sector:
    get:
      consumes:
      - application/json
      description: Compare company metrics with sector averages and peers
      parameters:
      - description: Company Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SectorComparisonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Get sector comparison for a company
      tags:
      - Analysis
  /api/v1/analysis/{symbol}/valuation:
    get:
      consumes:
      - application/json
      description: Get detailed valuation analysis with fair value estimates
      parameters:
      - description: Company Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.ValuationAnalysisResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Get valuation analysis for a company
      tags:
      - Analysis
  /api/v1/collect/{symbol}:
    post:
      consumes:
      - application/json
      description: Manually trigger data collection from Alpha Vantage for a specific
        company
      parameters:
      - description: Company Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Trigger data collection for a company
      tags:
      - Data Collection
  /api/v1/screening:
    post:
      consumes:
      - application/json
      description: Filter companies based on valuation, profitability, and financial
        health criteria
      parameters:
      - description: Screening criteria
        in: body
        name: criteria
        required: true
        schema:
          $ref: '#/definitions/services.ScreeningCriteria'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.ScreeningResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Screen companies based on fundamental criteria
      tags:
      - Screening
  /api/v1/screening/templates:
    get:
      consumes:
      - application/json
      description: Get predefined screening templates for common investment strategies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.ScreeningTemplatesResponse'
      summary: Get screening templates
      tags:
      - Screening
swagger: "2.0"